Remove DefaultMinimalResolutionResult
Remove in favor of chaning the interface to a concrete implementation. We no longer need multiple implementations of this type.
