+++ platforms/software/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/ivyservice/resolveengine/graph/results/DefaultVisitedGraphResultsTest.groovy
@@ -16,10 +16,8 @@
 
 package org.gradle.api.internal.artifacts.ivyservice.resolveengine.graph.results
 
-
 import org.gradle.api.artifacts.ResolveException
 import org.gradle.api.artifacts.UnresolvedDependency
-import org.gradle.api.internal.artifacts.result.DefaultMinimalResolutionResult
 import org.gradle.api.internal.artifacts.result.MinimalResolutionResult
 import org.gradle.api.internal.attributes.ImmutableAttributes
 import spock.lang.Specification
@@ -31,7 +29,7 @@ import java.util.function.Supplier
  */
 class DefaultVisitedGraphResultsTest extends Specification {
 
-    MinimalResolutionResult resolutionResult = new DefaultMinimalResolutionResult(Mock(Supplier), ImmutableAttributes.EMPTY)
+    MinimalResolutionResult resolutionResult = new MinimalResolutionResult(Mock(Supplier), ImmutableAttributes.EMPTY)
 
     def "hasResolutionFailure returns true if there is a failure"() {
         given:
